[anx] Move ANXGenerator to xchange-anx/src/test and remove the generators module, move ANX currencies and pairs to respective core classes,
