+++ pom.xml
@@ -123,7 +123,6 @@
 		<module>xchange-taurus</module>
 		<module>xchange-vircurex</module>
 		<module>xchange-yacuna</module>
-		<module>xchange-generators</module>
 	</modules>
 
 	<repositories>
