+++ xchange-anx/src/main/java/com/xeiam/xchange/anx/v2/service/polling/ANXBasePollingService.java
@@ -20,7 +20,6 @@ public class ANXBasePollingService extends BaseExchangeService implements BasePo
    * Constructor
    *
    * @param exchange
-   * @param nonceFactory
    */
   public ANXBasePollingService(Exchange exchange) {
