+++ xchange-core/src/main/java/com/xeiam/xchange/currency/CurrencyPair.java
@@ -112,6 +112,8 @@ public class CurrencyPair implements Comparable<CurrencyPair>{
   // BTC
   public static final CurrencyPair BTC_XDC = new CurrencyPair(Currencies.BTC, Currencies.XDC);
   public static final CurrencyPair BTC_PPC = new CurrencyPair(Currencies.BTC, Currencies.PPC);
+  public static final CurrencyPair STR_BTC = new CurrencyPair(Currencies.STR, Currencies.BTC);
+  public static final CurrencyPair XRP_BTC = new CurrencyPair(Currencies.XRP, Currencies.BTC);
 
   // LTC
   public static final CurrencyPair LTC_HKD = new CurrencyPair(Currencies.LTC, Currencies.HKD);
